Using optional sudo-specific password.
In certain server environments, such as secure server environments that use RSA
SecurID for server access, the password used with sudo can be different than
the password used for ssh authentication.  In these situations, it would be
convenient to be able to specify a sudo-specific password in the Capistrano
configuration in order to successfully use sudo with Capistrano.